+++ modules/core/src/main/java/org/rapidpm/microservice/optionals/header/Header.java
@@ -1,0 +1,13 @@
+package org.rapidpm.microservice.optionals.header;
+
+import java.lang.annotation.*;
+
+/**
+ * Created by svenruppert on 04.11.15.
+ */
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface Header {
+  int order() default Integer.MAX_VALUE;
+}