+++ modules/core/src/main/java/org/rapidpm/microservice/optionals/header/HeaderInfo.java
@@ -1,0 +1,9 @@
+package org.rapidpm.microservice.optionals.header;
+
+/**
+ * Created by svenruppert on 04.11.15.
+ */
+@FunctionalInterface
+public interface HeaderInfo {
+  String createHeaderInfo();
+}