+++ modules/core/src/test/java/junit/org/rapidpm/microservice/demo/FullTest.java
@@ -4,6 +4,7 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.rapidpm.ddi.DI;
 import org.rapidpm.microservice.Main;
 
 import java.io.BufferedReader;
@@ -23,6 +24,9 @@ public class FullTest {
 
   @Before
   public void setUp() throws Exception {
+    DI.clearReflectionModel();
+    DI.activatePackages("org.rapidpm");
+    DI.activatePackages(this.getClass().getPackage().getName());
     Main.deploy();
   }