+++ modules/optionals/cli/src/main/java/org/rapidpm/microservice/optionals/cli/CmdLineStartupAction.java
@@ -1,0 +1,14 @@
+package org.rapidpm.microservice.optionals.cli;
+
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+
+import java.util.List;
+
+/**
+ * Created by b.bosch on 18.11.2015.
+ */
+public interface CmdLineStartupAction {
+    List<Option> getOptions();
+    void execute(CommandLine cmdLine);
+}