+++ modules/optionals/cli/src/test/java/junit/org/rapidpm/microservice/optionals/cli/MainTest002.java
@@ -6,46 +6,48 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.rapidpm.ddi.DI;
 import org.rapidpm.microservice.Main;
-import org.rapidpm.microservice.optionals.cli.CmdLineSingleton;
+import org.rapidpm.microservice.optionals.cli.CmdLineParser;
+import org.rapidpm.microservice.optionals.cli.CmdLineStartupAction;
 
 import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 
 /**
  * Created by svenruppert on 31.08.15.
  */
-public class MainTest002 {
-
-  @Before
-  public void setUp() throws Exception {
-    final Optional<String[]> args = Optional.of(new String[]{"-i /opt/hoppel/poppel/xx.properties"});
-    Main.deploy(args);
-  }
-
-  @After
-  public void tearDown() throws Exception {
-    Main.stop();
-  }
+public class MainTest002 extends BaseCmdlineTest {
 
   @Test
   public void test001() throws Exception {
-
-    final CommandLine cmd = CmdLineSingleton.getInstance().getCommandLine().get();
-    Assert.assertNotNull(cmd);
-    Assert.assertTrue(cmd.hasOption("i"));
-
-    final String optionValue = cmd.getOptionValue("i");
-    Assert.assertEquals(optionValue.trim(), "/opt/hoppel/poppel/xx.properties");
-
+    final Optional<String[]> args = Optional.of(new String[]{"-i /opt/hoppel/poppel/xx.properties"});
+    Main.deploy(args);
+    Assert.assertTrue(DefaultCmdLineOptions.wasproperlyExecuted);
+    Main.stop();
   }
 
-  public static class DefaultCmdLineOptions implements Main.MainStartupAction {
+  public static class DefaultCmdLineOptions implements CmdLineStartupAction {
+
+    public static boolean wasproperlyExecuted = false;
 
     @Override
-    public void execute(Optional<String[]> args) {
+    public List<Option> getOptions() {
+      return Arrays.asList(new Option("i", "inifile", true, "some usefull stuff"));
+    }
+
+    @Override
+    public void execute(CommandLine cmdLine) {
       System.out.println("DefaultCmdLineOptions = " + LocalDateTime.now());
-      CmdLineSingleton.getInstance().addCmdLineOption(new Option("i", "inifile", true, "some usefull stuff"));
+
+      if(cmdLine.hasOption("i") &&  "/opt/hoppel/poppel/xx.properties".equals(cmdLine.getOptionValue("i").trim())){
+        wasproperlyExecuted = true;
+      }
+      else {
+        wasproperlyExecuted = false;
+      }
     }
   }