+++ release-notes.md
@@ -1,4 +1,15 @@
 # Release notes for RapidPM Microservices
 
-# 1.0
-Initial release
+# 0.0.3
+* custom CommandLineOptions
+* print all active ressources (urls) at startup
+
+
+# 0.0.4
+* added Header Infos
+* fixed bug which used the port for servlets even if there where none
+
+
+
+
+